AG-16708-set-data-value-in-batch
