+++ documentation/ag-grid-docs/src/content/api-documentation/grid-events/events.json
@@ -206,6 +206,18 @@
                 "url": "./cell-editing-full-row/"
             }
         },
+        "batchEditingStarted": {
+            "more": {
+                "name": "Batch Editing",
+                "url": "./cell-editing-batch/#batch-editing-lifecycle"
+            }
+        },
+        "batchEditingStopped": {
+            "more": {
+                "name": "Batch Editing",
+                "url": "./cell-editing-batch/#batch-editing-lifecycle"
+            }
+        },
         "undoStarted": {
             "more": {
                 "name": "Undo / Redo Events",
